Move integration list to separate file
